requirements.txt
Flask==3.0.0
flask_restful==0.3.10
SQLAlchemy==2.0.25
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy<1.24
tensorflow==2.16.1
pillow==10.3.0
